+++ messagebox.c
@@ -21,7 +21,7 @@
 
 #include "messagebox.h"
 
-static long appid;
+static int appid;
 
 static void close_callback(mtk_event *e, void *arg)
 {
